Several Updates
Updated the "slash" in paths the script outputs based on operating system. Some people found it annoying to have them mixed.

Removed regex dependency and replaced it with re so no external modules are required.

Added extra instructions when running the script. Some people thought you could give the script a URL when a PATH on drive is required.

Updated readme to show this instruction as well as compiling instruction.

Added compile.sh for Linux users.